backend/requirements.txt
# Core FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/ML and embeddings
openai==1.3.9
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.5

# Vector database
chromadb==0.4.18

# PDF processing
pdfplumber==0.10.3
PyPDF2==3.0.1

# Text processing
tiktoken==0.5.2

# Caching (optional)
redis==5.0.1

# Environment management
python-dotenv==1.0.0

# Utilities
requests==2.31.0
pydantic==2.5.0
typing-extensions==4.8.0

# Dev/testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging
structlog==23.2.0
